# ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Multi-Agent API",
    "description": "API with listing and markets agents",
    "version": "0.1.0"
  },
  "paths": {
...
    "/portal/profile": {
      "get": {
        "tags": [
          "portal"
        ],
        "summary": "Read Profile",
        "operationId": "read_profile_portal_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileOut"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "portal"
        ],
        "summary": "Update Profile",
        "operationId": "update_profile_portal_profile_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfilePatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/portal/tours": {
      "get": {
        "tags": [
          "portal"
        ],
        "summary": "Read Tours",
        "operationId": "read_tours_portal_tours_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "portal"
        ],
        "summary": "Sync Tours",
        "operationId": "sync_tours_portal_tours_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToursStatePayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/portal/move-in": {
      "get": {
        "tags": [
          "portal"
        ],
        "summary": "Read Move In",
        "operationId": "read_move_in_portal_move_in_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "portal"
        ],
        "summary": "Sync Move In",
        "operationId": "sync_move_in_portal_move_in_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveInStatePayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/portal/roommates": {
      "get": {
        "tags": [
          "portal"
        ],
        "summary": "Read Roommates",
        "operationId": "read_roommates_portal_roommates_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "portal"
        ],
        "summary": "Sync Roommates",
        "operationId": "sync_roommates_portal_roommates_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoommateStatePayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tours": {
      "get": {
        "tags": [
          "tours"
        ],
        "summary": "Read Tours",
        "operationId": "read_tours_tours_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "from_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From Date"
            }
          },
          {
            "name": "to_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To Date"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "created_at_desc",
              "title": "Sort"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToursListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "tours"
        ],
        "summary": "Create Tour Route",
        "operationId": "create_tour_route_tours_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TourCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/roommates/profile": {
      "get": {
        "tags": [
          "roommates"
        ],
        "summary": "Read My Roommate Profile",
        "operationId": "read_my_roommate_profile_roommates_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyRoommateProfileOut"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "roommates"
        ],
        "summary": "Patch My Roommate Profile",
        "operationId": "patch_my_roommate_profile_roommates_profile_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MyRoommateProfilePatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyRoommateProfileOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/listings/chat": {
      "post": {
        "summary": "Listing Chat",
        "operationId": "listing_chat_listings_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatRequest": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ChatAPIMessage"
            },
            "type": "array",
            "title": "Messages"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChatData"
              },
              {
                "type": "null"
              }
            ],
            "description": "The data of the chat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "messages"
        ],
        "title": "ChatRequest",
        "description": "The request to the chat API."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MoveInStatePayload": {
        "properties": {
          "plan": {
            "$ref": "#/components/schemas/MoveInPlanPayload"
          },
          "orders": {
            "items": {
              "$ref": "#/components/schemas/VendorOrderPayload"
            },
            "type": "array",
            "title": "Orders"
          },
          "checklist": {
            "items": {
              "$ref": "#/components/schemas/ChecklistItemPayload"
            },
            "type": "array",
            "title": "Checklist"
          }
        },
        "type": "object",
        "title": "MoveInStatePayload"
      },
...
      "ProfileOut": {
        "properties": {
          "has_current_lease": {
            "type": "boolean",
            "title": "Has Current Lease",
            "default": false
          },
          "search_trigger": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Search Trigger"
          },
          "trigger_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trigger Reason"
          },
          "move_timeline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Move Timeline"
          },
          "current_city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current City"
          },
          "work_location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Work Location"
          },
          "preferred_cities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Preferred Cities"
          },
          "neighbourhood_priorities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Neighbourhood Priorities"
          },
          "dealbreakers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Dealbreakers"
          },
          "max_monthly_rent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Monthly Rent"
          },
          "credit_score_range": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credit Score Range"
          },
          "living_arrangement": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Living Arrangement"
          },
          "roommate_search_enabled": {
            "type": "boolean",
            "title": "Roommate Search Enabled",
            "default": false
          },
          "bedrooms_needed": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bedrooms Needed"
          },
          "has_pets": {
            "type": "boolean",
            "title": "Has Pets",
            "default": false
          },
          "pet_details": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pet Details"
          },
          "journey_stage_override": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Journey Stage Override"
          },
          "onboarding_completed": {
            "type": "boolean",
            "title": "Onboarding Completed",
            "default": false
          },
          "onboarding_step": {
            "type": "integer",
            "title": "Onboarding Step",
            "default": 0
          },
          "last_updated": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Updated"
          }
        },
        "type": "object",
        "title": "ProfileOut"
      },
      "ProfilePatch": {
        "properties": {
          "has_current_lease": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Has Current Lease"
          },
          "search_trigger": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Search Trigger"
          },
          "trigger_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trigger Reason"
          },
          "move_timeline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Move Timeline"
          },
          "current_city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current City"
          },
          "work_location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Work Location"
          },
          "preferred_cities": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Cities"
          },
          "neighbourhood_priorities": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Neighbourhood Priorities"
          },
          "dealbreakers": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dealbreakers"
          },
          "max_monthly_rent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Monthly Rent"
          },
          "credit_score_range": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credit Score Range"
          },
          "living_arrangement": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Living Arrangement"
          },
          "roommate_search_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roommate Search Enabled"
          },
          "bedrooms_needed": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bedrooms Needed"
          },
          "has_pets": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Has Pets"
          },
          "pet_details": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pet Details"
          },
          "journey_stage_override": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Journey Stage Override"
          },
          "onboarding_completed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Onboarding Completed"
          },
          "onboarding_step": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Onboarding Step"
          }
        },
        "type": "object",
        "title": "ProfilePatch"
      },
...
      "RoommateStatePayload": {
        "properties": {
          "my_profile": {
            "$ref": "#/components/schemas/MyRoommateProfilePayload"
          },
          "connections": {
            "items": {
              "$ref": "#/components/schemas/RoommateConnectionPayload"
            },
            "type": "array",
            "title": "Connections"
          }
        },
        "type": "object",
        "title": "RoommateStatePayload"
      },
...
      "TourCreate": {
        "properties": {
          "property": {
            "$ref": "#/components/schemas/tours__schemas__TourPropertyPayload"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "saved"
          },
          "scheduled_date": {
            "type": "string",
            "title": "Scheduled Date",
            "default": ""
          },
          "scheduled_time": {
            "type": "string",
            "title": "Scheduled Time",
            "default": ""
          },
          "note": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TourNotePayload"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "property"
        ],
        "title": "TourCreate"
      },
...
      "TourResponse": {
        "properties": {
          "tour": {
            "$ref": "#/components/schemas/TourPayload-Output"
          }
        },
        "type": "object",
        "required": [
          "tour"
        ],
        "title": "TourResponse"
      },
...
      "ToursListResponse": {
        "properties": {
          "tours": {
            "items": {
              "$ref": "#/components/schemas/TourPayload-Output"
            },
            "type": "array",
            "title": "Tours"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "total"
        ],
        "title": "ToursListResponse"
      },
      "ToursStatePayload": {
        "properties": {
          "tours": {
            "items": {
              "$ref": "#/components/schemas/TourPayload-Input"
            },
            "type": "array",
            "title": "Tours"
          }
        },
        "type": "object",
        "title": "ToursStatePayload"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}